Se dividen las páginas en nuevos componentes para poder agregar el metadata evitando utilizar use client
