+++ main/app/layout.jsx
@@ -8,7 +8,7 @@ import Terms from "@components/Terms";
 import { Analytics } from "@vercel/analytics/react";
 
 export const metadata = {
-  title: "EncuentraLoFácilCR: Descubre, Publica y Encuentra Autos, Viviendas, Abarrotes y Servicios en Costa Rica",
+  title: "EncuéntraLo Fácil CR: Descubre, Publica y Encuentra Autos, Viviendas, Abarrotes y Servicios en Costa Rica",
   description: "Explora Encuentralo Fácil CR, tu destino en línea en Costa Rica para descubrir y publicar una amplia variedad de bienes y servicios. Encuentra fácilmente autos, viviendas, productos y servicios. Publica tus propios productos y servicios con facilidad. ¡Haz tus búsquedas más eficientes y disfruta de la comodidad en EncuentraLoFácilCR!",
 };
