+++ main/app/unauthorized/page.jsx
@@ -1,5 +1,10 @@
 import React from 'react'
 
+export const metadata = {
+  title: "EncuéntraLo Fácil CR: Acceso no autorizado",
+  description: "Encuentralo Fácil CR, acceso no autorizado.",
+};
+
 const Unauthorized = () => {
   return (
     <div>401 No está autorizado a ver esta página</div>
